+++ src/Opulence/Debug/Exceptions/Handlers/Http/templates/DevelopmentExceptionPage.php
@@ -14,10 +14,24 @@
             {
                 text-decoration: none;
             }
+
+            ul.errors
+            {
+                list-style: disc inside none;
+            }
         </style>
     </head>
     <body>
         <h2>Something went wrong</h2>
-        <?php echo htmlentities($ex->getMessage(), ENT_QUOTES, "UTF-8"); ?>
+        <ul class="errors">
+            <li>
+                <pre><?php echo htmlentities($ex->getMessage(), ENT_QUOTES, "UTF-8"); ?></pre>
+            </li>
+            <?php while ($ex = $ex->getPrevious()): ?>
+                <li>
+                    <pre><?php echo htmlentities($ex->getMessage(), ENT_QUOTES, "UTF-8"); ?></pre>
+                </li>
+            <?php endwhile; ?>
+        </ul>
     </body>
 </html>