+++ src/Opulence/Routing/Routes/Compilers/Parsers/IParser.php
@@ -8,9 +8,9 @@
  */
 namespace Opulence\Routing\Routes\Compilers\Parsers;
 
-use Opulence\Routing\Routes\Route;
 use Opulence\Routing\RouteException;
 use Opulence\Routing\Routes\ParsedRoute;
+use Opulence\Routing\Routes\Route;
 
 /**
  * Defines the interface for route parsers to implement